+++ src/model/playing.rs
@@ -4,7 +4,6 @@ use serde::{Deserialize, Serialize};
 
 use super::context::Context;
 use super::track::FullTrack;
-use super::track::SimplifiedTrack;
 /// Current playing track
 /// [get the users currently playing track](https://developer.spotify.com/web-api/get-the-users-currently-playing-track/)
 #[derive(Clone, Debug, Serialize, Deserialize)]
@@ -20,7 +19,7 @@ pub struct Playing {
 /// [play history object](https://developer.spotify.com/web-api/object-model/#play-history-object)
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct PlayHistory {
-    pub track: SimplifiedTrack,
+    pub track: FullTrack,
     pub played_at: DateTime<Utc>,
     pub context: Option<Context>,
 }